Added audio playback of cues using awedio.
This change adds a media player thread that waits
on filepaths to be sent from the main thread when
a sound cue button is pressed. Also added is a
cache that maps filepaths to awedio `MemorySound`s
so that file I/O does not have to take place each
button press. 9 test audio cue files are included.
closes #2